Moving some more tests.
