+++ job/src/test/java/org/jenkinsci/plugins/workflow/WorkflowRunTest.java
@@ -84,7 +84,7 @@ public class WorkflowRunTest {
 
     @Test public void parameters() throws Exception {
         WorkflowJob p = r.jenkins.createProject(WorkflowJob.class, "p");
-        p.setDefinition(new CpsFlowDefinition("node {sh('echo param=' + PARAM)}",true));
+        p.setDefinition(new CpsFlowDefinition("echo \"param=${PARAM}\"",true));
         p.addProperty(new ParametersDefinitionProperty(new StringParameterDefinition("PARAM", null)));
         WorkflowRun b = r.assertBuildStatusSuccess(p.scheduleBuild2(0, new ParametersAction(new StringParameterValue("PARAM", "value"))));
         r.assertLogContains("param=value", b);
@@ -187,14 +187,6 @@ public class WorkflowRunTest {
         assertEquals("devel", pendingScript.getContext().getUser());
         ScriptApproval.get().approveScript(pendingScript.getHash());
         r.assertLogContains("hello", r.assertBuildStatusSuccess(p.scheduleBuild2(0)));
-    }
-
-    @Test @Issue("JENKINS-25630")
-    public void contextInjectionOfSubParameters() throws Exception {
-        // see SubtypeInjectingStep
-        WorkflowJob p = r.jenkins.createProject(WorkflowJob.class, "p");
-        p.setDefinition(new CpsFlowDefinition("node('master') { injectSubtypesAsContext() }"));
-        r.assertBuildStatusSuccess(p.scheduleBuild2(0));
     }
 
     @Test @Issue("JENKINS-29221")
